+++ packages/hub/src/scope.ts
@@ -7,8 +7,8 @@ import { SyncPromise } from '@sentry/utils/syncpromise';
 /**
  * Event processors are used to change the event before it will be send.
  * We strongly advise to make this function sync.
- * Returning a Promise<SentryEvent | null> works just fine, just be sure what you are doing.
- * Eventprosccing will be stopped until your Promise is resolved.
+ * Returning a Promise<SentryEvent | null> will work just fine, but better be sure that you know what you are doing.
+ * Event processing will be deferred until your Promise is resolved.
  */
 export type EventProcessor = (
   event: SentryEvent,