+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- [loader] fix: Init and replay all calls to `Sentry.*` first before invoking all registered callbacks
 - [browser] fix: DOMError and DOMException should be error level events
 - [utils] ref: Update wrap method to hide internal sentry flags
 - [utils] fix: Make internal Sentry flags non-enumerable in fill util