+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.5.4
+
 - [browser] fix: `DOMError` and `DOMException` should be error level events
 - [browser] ref: Log error if Ember/Vue instances are not provided
 - [utils] fix: Dont mutate original input in `decycle` util function